# PolicyRepository: index and replace rules by resource.