--- a/pkg/policy/k8s/cilium_network_policy.go
+++ b/pkg/policy/k8s/cilium_network_policy.go
@@ -138,10 +138,10 @@ func (p *policyWatcher) upsertCiliumNetworkPolicyV2(cnp *types.SlimCNP, initialR
 	rules, policyImportErr := cnp.Parse()
 	if policyImportErr == nil {
 		_, policyImportErr = p.policyManager.PolicyAdd(rules, &policy.AddOptions{
-			ReplaceWithLabels:   cnp.GetIdentityLabels(),
 			Source:              source.CustomResource,
 			ProcessingStartTime: initialRecvTime,
 			Resource:            resourceID,
+			ReplaceByResource:   true,
 		})
 	}
 
@@ -163,9 +163,10 @@ func (p *policyWatcher) deleteCiliumNetworkPolicyV2(cnp *types.SlimCNP, resource
 
 	scopedLog.Debug("Deleting CiliumNetworkPolicy")
 
-	_, err := p.policyManager.PolicyDelete(cnp.GetIdentityLabels(), &policy.DeleteOptions{
-		Source:   source.CustomResource,
-		Resource: resourceID,
+	_, err := p.policyManager.PolicyDelete(nil, &policy.DeleteOptions{
+		Source:           source.CustomResource,
+		Resource:         resourceID,
+		DeleteByResource: true,
 	})
 	if err == nil {
 		scopedLog.Info("Deleted CiliumNetworkPolicy")