@squeed 2024-05-27
Tagging @aanm for review too; you understand this bit of code somewhat.