@squeed 2024-05-30
Tests caught a flake -- just some test code that expected a certain ordering. Fixed.