@aanm 2024-05-30
> > Do we have any performance benchmark numbers that can tell us how much this improves the startup time?
> 
> @christarazi I threw together a quick benchmark:
> 
> ```
> $ benchstat bench-before.txt bench-after.txt 
> goos: linux
> goarch: amd64
> pkg: github.com/cilium/cilium/pkg/policy
> cpu: 12th Gen Intel(R) Core(TM) i7-1250U
>             │ bench-before.txt │           bench-after.txt            │
>             │      sec/op      │    sec/op     vs base                │
> AddRules-12      241.740µ ± 7%   1.538µ ± 49%  -99.36% (p=0.000 n=10)
> ```
> 
> Good enough for me :-)

I also ran a local benchmark to compare the PR against current main and I didn't see any regressions.